% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONDITIONS}
\alias{CONDITIONS}
\title{CONDITIONS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#individualConditions>
}
\usage{
CONDITIONS
}
\description{
CONDITIONS
}
\keyword{datasets}
