% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COUNTRIES}
\alias{COUNTRIES}
\title{COUNTRIES}
\format{
\describe{
  \item{CODE}{The code}
  \item{CODE_ISO2}{The ISO2 country code - when available}
  \item{CODE_ISO3}{The ISO3 country code - when available}
  \item{LIST_NAME_EN}{The English name (when appearing in a list)}
  \item{LIST_NAME_FR}{The French name (when appearing in a list)}
  \item{FULL_NAME_EN}{The full English name}
  \item{FULL_NAME_FR}{The full French name}
  \item{SHORT_NAME_EN}{The short English name}
  \item{SHORT_NAME_FR}{The short French name}
  \item{NOCS_CATEGORY}{The category within the FAO NOCS database}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/admin/#countries>
}
\usage{
COUNTRIES
}
\description{
COUNTRIES
}
\keyword{datasets}
