% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_COVERAGE_TYPES}
\alias{DATA_COVERAGE_TYPES}
\title{DATA_COVERAGE_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/data/#coverageTypes>
}
\usage{
DATA_COVERAGE_TYPES
}
\description{
DATA_COVERAGE_TYPES
}
\keyword{datasets}
