% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_ESTIMATIONS}
\alias{DATA_ESTIMATIONS}
\title{DATA_ESTIMATIONS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
DATA_ESTIMATIONS
}
\description{
DATA_ESTIMATIONS
}
\keyword{datasets}
