% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA_PROCESSINGS}
\alias{DATA_PROCESSINGS}
\title{DATA_PROCESSINGS}
\format{
\describe{
  \item{DATASET_CODE}{The dataset code}
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/data/#processings>
}
\usage{
DATA_PROCESSINGS
}
\description{
DATA_PROCESSINGS
}
\keyword{datasets}
