% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DISCARD_REASONS}
\alias{DISCARD_REASONS}
\title{DISCARD_REASONS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#discardReasons>
}
\usage{
DISCARD_REASONS
}
\description{
DISCARD_REASONS
}
\keyword{datasets}
