% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EEZ_TO_IOTC_MAIN_AREAS}
\alias{EEZ_TO_IOTC_MAIN_AREAS}
\title{EEZ_TO_IOTC_MAIN_AREAS}
\format{
\describe{
  \item{FLAG_CODE}{The flag code of a coastal state}
  \item{MAIN_IOTC_AREA_CODE}{The main IOTC area code overlapping with the EEZ of the coastal state}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
EEZ_TO_IOTC_MAIN_AREAS
}
\description{
EEZ_TO_IOTC_MAIN_AREAS
}
\keyword{datasets}
