% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EFFORT_UNITS}
\alias{EFFORT_UNITS}
\title{EFFORT_UNITS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#effortUnits>
}
\usage{
EFFORT_UNITS
}
\description{
EFFORT_UNITS
}
\keyword{datasets}
