% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAD_ACTIVITY_TYPES}
\alias{FAD_ACTIVITY_TYPES}
\title{FAD_ACTIVITY_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{IS_ON_DFAD}{Whether or not the activity is on DFAD}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#FADactivityTypes>
}
\usage{
FAD_ACTIVITY_TYPES
}
\description{
FAD_ACTIVITY_TYPES
}
\keyword{datasets}
