% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAD_OWNERSHIP_TYPES}
\alias{FAD_OWNERSHIP_TYPES}
\title{FAD_OWNERSHIP_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#FADownerships>
}
\usage{
FAD_OWNERSHIP_TYPES
}
\description{
FAD_OWNERSHIP_TYPES
}
\keyword{datasets}
