% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FATES}
\alias{FATES}
\title{FATES}
\format{
\describe{
  \item{TYPE_OF_FATE_CODE}{The type of fate - i.e., retained / discarded}
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#fates>
}
\usage{
FATES
}
\description{
FATES
}
\keyword{datasets}
