% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOB_ACTIVITY_TYPES}
\alias{FOB_ACTIVITY_TYPES}
\title{FOB_ACTIVITY_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{DFOB}{Whether or not the activity is on drifting FOBs / FADs}
  \item{AFOB}{Whether or not the activity is on anchored FOBs / FADs}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#FOBactivityTypes>
}
\usage{
FOB_ACTIVITY_TYPES
}
\description{
FOB_ACTIVITY_TYPES
}
\keyword{datasets}
