% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOB_TYPES}
\alias{FOB_TYPES}
\title{FOB_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{DFOB}{Whether or not the FOB type identifies a drifting FOB / FAD}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#FOBtypes>
}
\usage{
FOB_TYPES
}
\description{
FOB_TYPES
}
\keyword{datasets}
