% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IO_GRIDS_05x05}
\alias{IO_GRIDS_05x05}
\title{IO_GRIDS_05x05}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/admin/#IOgrids5x5>
}
\usage{
IO_GRIDS_05x05
}
\description{
IO_GRIDS_05x05
}
\keyword{datasets}
