% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IOTC_GRIDS_10x10}
\alias{IOTC_GRIDS_10x10}
\title{IOTC_GRIDS_10x10}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{OCEAN_AREA_KM2}{The ocean area in km2}
  \item{OCEAN_AREA_IO_KM2}{The ocean area in km2 (limited to the area that falls within the Indian Ocean)}
  \item{OCEAN_AREA_IOTC_KM2}{The ocean area in km2 (limited to the area that falls within the IOTC competence area)}
  \item{CENTER_LAT}{The latitude of the centroid}
  \item{CENTER_LON}{The longitude of the centroid}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/admin/#IOTCgrids10x10>
}
\usage{
IOTC_GRIDS_10x10
}
\description{
IOTC_GRIDS_10x10
}
\keyword{datasets}
