% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IOTC_GRIDS_AR}
\alias{IOTC_GRIDS_AR}
\title{IOTC_GRIDS_AR}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
IOTC_GRIDS_AR
}
\description{
IOTC_GRIDS_AR
}
\keyword{datasets}
