% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IOTC_MAIN_AREAS}
\alias{IOTC_MAIN_AREAS}
\title{IOTC_MAIN_AREAS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/admin/#IOTCareasMain>
}
\usage{
IOTC_MAIN_AREAS
}
\description{
IOTC_MAIN_AREAS
}
\keyword{datasets}
