% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LEGACY_FLEETS}
\alias{LEGACY_FLEETS}
\title{LEGACY_FLEETS}
\format{
\describe{
  \item{FLAG_CODE}{The flag country code}
  \item{REPORTING_ENTITY_CODE}{The reporting entity code}
  \item{FLEET_CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
LEGACY_FLEETS
}
\description{
LEGACY_FLEETS
}
\keyword{datasets}
