% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MECHANIZATION_TYPES}
\alias{MECHANIZATION_TYPES}
\title{MECHANIZATION_TYPES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{IS_MECHANIZED}{Whether or not the mechanization type refers to mechanized boats}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#mechanizationTypes>
}
\usage{
MECHANIZATION_TYPES
}
\description{
MECHANIZATION_TYPES
}
\keyword{datasets}
