% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RETAIN_REASONS}
\alias{RETAIN_REASONS}
\title{RETAIN_REASONS}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#retainReasons>
}
\usage{
RETAIN_REASONS
}
\description{
RETAIN_REASONS
}
\keyword{datasets}
