% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEX}
\alias{SEX}
\title{SEX}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{IS_DETERMINED}{Whether or not the sex is determined}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#sex>
}
\usage{
SEX
}
\description{
SEX
}
\keyword{datasets}
