% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TYPES_OF_FATE}
\alias{TYPES_OF_FATE}
\title{TYPES_OF_FATE}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
TYPES_OF_FATE
}
\description{
TYPES_OF_FATE
}
\keyword{datasets}
