% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TYPES_OF_MEASUREMENT}
\alias{TYPES_OF_MEASUREMENT}
\title{TYPES_OF_MEASUREMENT}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{DEFAULT_UNIT}{The default unit for the given type of measurement}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/biological/#typesOfMeasurement>
}
\usage{
TYPES_OF_MEASUREMENT
}
\description{
TYPES_OF_MEASUREMENT
}
\keyword{datasets}
