% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FISHERIES}
\alias{FISHERIES}
\title{FISHERIES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{FISHERY_CATEGORY_CODE}{The fishery category code (as per Res. 15/02)}
  \item{FISHERY_CATEGORY_NAME_EN}{The fishery category English name (as per Res. 15/02)}
  \item{FISHERY_CATEGORY_NAME)FR}{The fishery category French name (as per Res. 15/02)}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/fisheries/#fisheries>
}
\usage{
FISHERIES
}
\description{
FISHERIES
}
\keyword{datasets}
