% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IOTC_MAIN_AREAS}
\alias{IOTC_MAIN_AREAS}
\title{IOTC_MAIN_AREAS
Reference codes modeling the two main areas in which the IOTC area of competence is split (western / eastern IOTC area, roughly coinciding with FAO areas F51 and F57).}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{OCEAN_AREA_KM2}{The ocean area in km2}
  \item{OCEAN_AREA_IO_KM2}{The ocean area in km2 (limited to the area that falls within the Indian Ocean)}
  \item{OCEAN_AREA_IOTC_KM2}{The ocean area in km2 (limited to the area that falls within the IOTC competence area)}
  \item{CENTER_LAT}{The latitude of the centroid}
  \item{CENTER_LON}{The longitude of the centroid}
}
}
\source{
<https://data.iotc.org/reference/latest/domain/admin/#IOTCareasMain>
}
\usage{
IOTC_MAIN_AREAS
}
\description{
IOTC_MAIN_AREAS
Reference codes modeling the two main areas in which the IOTC area of competence is split (western / eastern IOTC area, roughly coinciding with FAO areas F51 and F57).
}
\keyword{datasets}
