% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LEGACY_FISHERIES}
\alias{LEGACY_FISHERIES}
\title{LEGACY_FISHERIES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{IS_AGGREGATE}{If the fishery is an aggregate}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
LEGACY_FISHERIES
}
\description{
LEGACY_FISHERIES
}
\keyword{datasets}
