% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LEGACY_SPECIES}
\alias{LEGACY_SPECIES}
\title{LEGACY_SPECIES}
\format{
\describe{
  \item{CODE}{The code}
  \item{NAME_EN}{The English name}
  \item{NAME_FR}{The French name}
  \item{NAME_SCIENTIFIC}{The scientific name}
  \item{IS_IOTC}{If the species belongs to the 16 IOTC ones}
  \item{IS_AGGREGATE}{If the species is an aggregate}
}
}
\source{
<https://data.iotc.org/reference/latest>
}
\usage{
LEGACY_SPECIES
}
\description{
LEGACY_SPECIES
}
\keyword{datasets}
